work with files added

+++ App/FileWork/FileWork.cs
@@ -3,35 +3,46 @@
 using System.Text.Json;
 public class FileWork
 {
+    Exception FileWasNotFound;
+
     private const string _directoryPath = "C://Khai";
-    private const string _infoFilePath = "C://Khai/ifo.json";
+    private const string _infoFilePath = "C://Khai/info.json";
    
-    private static void CheckIsFileExist()
+    private static int CheckIsFileExist()
     {
         if (!Directory.Exists(_directoryPath))
-            throw new Exception("Directory is not existing");
+            return -1;
+            //throw new Exception("Directory is not existing");
 
         if (!File.Exists(_infoFilePath))
-            throw new Exception("File is not existing");
+            return -2;
+            //throw new Exception("File is not existing");
+
+        return 0;
     }
 
     public static void CreateAFile()
     {
         if (!Directory.Exists(_directoryPath))
             Directory.CreateDirectory(_directoryPath);
-            
-        if(!File.Exists(_infoFilePath))
-            File.Create(_infoFilePath);
+
+        if (!File.Exists(_infoFilePath))
+        {
+            var SchFile = File.Create(_infoFilePath);
+            SchFile.Close();
+        }         
     }
 
     public static void SaveSchduleToFile(WeekSchedule weekSchedule)
     {
-        try
-        {
-            CheckIsFileExist();
-            var json = JsonSerializer.Serialize(weekSchedule);
-            File.WriteAllText(_infoFilePath, json);
-        } catch { }
+
+            if (CheckIsFileExist() == 0)
+            {
+                var json = JsonSerializer.Serialize(weekSchedule);
+                File.WriteAllText(_infoFilePath, json);
+            }
+            else if(CheckIsFileExist() == -1) throw new DirectoryDoesNotExistException("File is not existing");
+            else throw new FileDoesNotExistException("Directory is not existing");
         
     }
 
@@ -52,7 +63,7 @@ public class FileWork
             var rowJSON = File.ReadAllText(_infoFilePath);
 
             if (rowJSON.Length == 0)
-                throw new Exception("Немає збереженого розкладу.");
+                throw new ScheduleWasNotFoundException("Немає збереженого розкладу.");
 
              readed = JsonSerializer.Deserialize<Khai.WeekSchedule>(rowJSON);
         } catch {
@@ -60,8 +71,32 @@ public class FileWork
         }
 
         if (readed == null)
-            throw new Exception("Не вдалося прочитати файл.");
+            throw new FileWasNotFoundException("Не вдалося прочитати файл.");
 
         return readed;
     }
+}
+
+ public class FileWasNotFoundException : Exception
+{
+    public FileWasNotFoundException(string message)
+        : base(message) { }
+}
+
+public class ScheduleWasNotFoundException : Exception
+{
+    public ScheduleWasNotFoundException(string message)
+        : base(message) { }
+}
+
+public class DirectoryDoesNotExistException : Exception
+{
+    public DirectoryDoesNotExistException(string message)
+        : base(message) { }
+}
+
+public class FileDoesNotExistException : Exception
+{
+    public FileDoesNotExistException(string message)
+        : base(message) { }
 }
